[#8] add: AppExceptionを導入し例外定義を集約
