+++ lib/data/api/weather_api.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'package:flutter_training/domain/exception/app_exception.dart';
 import 'package:yumemi_weather/yumemi_weather.dart';
 
 class WeatherAPI {
@@ -27,17 +28,4 @@ class WeatherAPI {
       'date': value.dateTime.toString(),
     };
   }
-}
-
-sealed class WeatherAPIError implements Exception {
-  const WeatherAPIError({required this.message});
-  final String message;
-}
-
-class InvalidParameter extends WeatherAPIError {
-  const InvalidParameter() : super(message: '無効な都道府県が指定されました');
-}
-
-class Unknown extends WeatherAPIError {
-  const Unknown() : super(message: '天気情報が取得できません');
 }
