+++ lib/domain/model/weather_info.dart
