+++ lib/domain/model/weather_info.freezed.dart
